// src/lib/api/gen/types/ListRaffles.ts
/**
* Generated by Kubb (https://kubb.dev/).
* Do not edit manually.
*/


export type ListRafflesPathParams = {
    /**
     * @type string
    */
    slug: string;
    /**
     * @type string
    */
    campaignSlug: string;
};

export const itemsStatusEnum5 = {
    open: "open",
    drawn: "drawn",
    cancelled: "cancelled"
} as const;

export type ItemsStatusEnum5Key = (typeof itemsStatusEnum5)[keyof typeof itemsStatusEnum5];

/**
 * @description Default Response
*/
export type ListRaffles200 = {
    /**
     * @type array
    */
    items: {
        /**
         * @type string
        */
        campaignSlug: string;
        /**
         * @minLength -9007199254740991
         * @maxLength 9007199254740991
         * @type integer
        */
        sequence: number;
        /**
         * @type string
        */
        title: string;
        /**
         * @type string
        */
        startsAt: string;
        /**
         * @type string
        */
        endsAt: string | null;
        /**
         * @minLength -9007199254740991
         * @maxLength 9007199254740991
         * @type integer
        */
        centsPerNumber: number;
        /**
         * @type string
        */
        drawAt: string | null;
        /**
         * @type string
        */
        status: ItemsStatusEnum5Key;
        /**
         * @type string
        */
        seed: string | null;
        /**
         * @type string
        */
        algorithm: string;
        /**
         * @type string
        */
        drawnAt: string | null;
        /**
         * @minLength -9007199254740991
         * @maxLength 9007199254740991
         * @type integer
        */
        totalEntries: number;
        /**
         * @minLength -9007199254740991
         * @maxLength 9007199254740991
         * @type integer
        */
        totalParticipants: number;
        /**
         * @type array
        */
        prizes: {
            /**
             * @minLength -9007199254740991
             * @maxLength 9007199254740991
             * @type integer
            */
            position: number;
            /**
             * @type string
            */
            title: string;
            /**
             * @type string
            */
            description: string | null;
            /**
             * @type array
            */
            images: string[];
            /**
             * @type array
            */
            imageUrls: string[];
            /**
             * @type object
            */
            winner: {
                /**
                 * @minLength -9007199254740991
                 * @maxLength 9007199254740991
                 * @type integer
                */
                number: number;
                /**
                 * @type string
                */
                participant: string;
            } | null;
        }[];
    }[];
};

export type ListRafflesQueryResponse = ListRaffles200;

export type ListRafflesQuery = {
    Response: ListRaffles200;
    PathParams: ListRafflesPathParams;
    Errors: any;
};